requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
gunicorn==20.1.0
SQLAlchemy==1.4
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
pytest==7.0.1
Werkzeug==2.2.2